resolve circular deps
