+++ packages/ast/types/src/docs/rpc-clients.d.ts
@@ -1,6 +1,5 @@
 import * as t from '@babel/types';
 import { ProtoRef, ProtoService, ProtoServiceMethod } from "@cosmology/types";
-import { ProtoStore } from "@cosmology/proto-parser";
 import { ProtoParseContext } from "../encoding";
 interface DocumentRpcClient {
     service: DocumentService;
@@ -13,6 +12,6 @@ interface DocumentService {
     svc: ProtoService;
     ref: ProtoRef;
 }
-export declare const documentRpcClients: (context: ProtoParseContext, myBase: string, store: ProtoStore) => DocumentRpcClient[];
-export declare const documentRpcClientsReadme: (context: ProtoParseContext, myBase: string, store: ProtoStore) => string;
+export declare const documentRpcClients: (context: ProtoParseContext, services: Record<string, ProtoRef[]>) => DocumentRpcClient[];
+export declare const documentRpcClientsReadme: (context: ProtoParseContext, services: Record<string, ProtoRef[]>) => string;
 export {};
