+++ packages/ast/types/src/plugins/plugins.d.ts
@@ -1,2 +1,1 @@
-import { TelescopeOptions, TelescopeOption } from '@cosmology/types';
-export declare const getPluginValue: (optionName: TelescopeOption | string, currentPkg: string, options: TelescopeOptions) => any;
+export { getPluginValue } from '@cosmology/utils';
