+++ packages/parser/types/lookup.d.ts
@@ -1,13 +1,7 @@
 import { ProtoStore } from './store';
 import { ProtoRef } from '@cosmology/types';
-export interface Lookup {
-    obj: any;
-    name: string;
-    import: string;
-    importType: string;
-    importedName: string;
-    package: string;
-}
+import { Lookup } from '@cosmology/types';
+export type { Lookup } from '@cosmology/types';
 export declare const recursiveLookup: (proto: any, name: string, scope?: string[], allowNested?: boolean) => any;
 export declare const importLookup: (store: ProtoStore, ref: ProtoRef, name: string) => {
     name: string;
