+++ packages/parser/types/utils.d.ts
@@ -1,28 +1,8 @@
 import { InterfaceTypeUrlMap, ProtoRef, ProtoRoot } from '@cosmology/types';
 import { ProtoStore } from './store';
+export { isRefExcluded, isRefIncluded, getObjectName } from '@cosmology/utils';
 export declare const lookupSymbolScopes: (name: string, relativeTo: string, list?: string[]) => any;
 export declare const createTypeUrlTypeMap: (store: ProtoStore, fromRef: ProtoRef) => InterfaceTypeUrlMap;
-/**
- * test if a proto ref is included by the operation.
- * @param ref a ProtoRef with proto file info and package.
- * @param include patterns(will be deprecated soon), packages, proto files to include
- * @returns
- */
-export declare const isRefIncluded: (ref: ProtoRef, include?: {
-    patterns?: string[];
-    packages?: string[];
-    protos?: string[];
-}) => boolean;
-/**
- * test if a proto ref is excluded from the operation.
- * @param ref a ProtoRef with proto file info and package.
- * @param exclude patterns(will be deprecated soon), packages, proto files to exclude
- * @returns
- */
-export declare const isRefExcluded: (ref: ProtoRef, exclude?: {
-    packages?: string[];
-    protos?: string[];
-}) => boolean;
 export declare const getPackageAndNestedFromStr: (type: string, pkg: string) => {
     nested: string;
     package: string;
@@ -30,7 +10,6 @@ export declare const getPackageAndNestedFromStr: (type: string, pkg: string) => 
 export declare const getServices: (root: ProtoRoot) => any[];
 export declare const getTypes: (root: ProtoRoot) => any[];
 export declare const getEnums: (root: ProtoRoot) => any[];
-export declare const getObjectName: (name: string, scope?: string[]) => string;
 export declare const SCALAR_TYPES: string[];
 export declare const instanceType: (obj: any) => {
     type: string;
