+++ packages/types/types/base.d.ts
@@ -1,3 +1,18 @@
+import { TelescopeOptions } from "./telescope";
+export interface IParseContext {
+    ref: ProtoRef;
+    options: TelescopeOptions;
+    imports: ImportUsage[];
+    utils: Record<string, boolean>;
+    addUtil: Function;
+    pluginValue(name: any): any;
+}
+export interface ImportUsage {
+    type: 'typeImport' | 'toJSONEnum' | 'fromJSONEnum';
+    name: string;
+    import: string;
+    importedAs?: string;
+}
 export interface ProtoEnum {
     type?: 'Enum';
     name?: string;
@@ -158,3 +173,19 @@ export type InterfaceTypeUrlMap = Record<string, TraverseTypeUrlRef[]>;
 export type TraversalSymbol = TraverseLocalSymbol & {
     ref: string;
 };
+export interface Lookup {
+    obj: any;
+    name: string;
+    import: string;
+    importType: string;
+    importedName: string;
+    package: string;
+}
+export interface IProtoStore {
+    requests: Record<string, ProtoServiceMethod>;
+    responses: Record<string, ProtoServiceMethod>;
+    _symbols: TraversalSymbol[];
+    findProto(filename: any): ProtoRef;
+    get(from: ProtoRef, name: string): Lookup;
+    getTypeUrlMap(ref: ProtoRef): any;
+}
