+++ packages/utils/src/index.ts
@@ -2,3 +2,4 @@ export * from './case';
 export * from './slugs';
 export * from './proto';
 export * from './utils';
+export * from './amino';
