+++ packages/utils/types/proto.d.ts
@@ -1,4 +1,4 @@
-import { ProtoRoot } from "@cosmology/types";
+import { ProtoRef, ProtoRoot } from "@cosmology/types";
 export declare const getNestedProto: (root: ProtoRoot) => any;
 export declare const getNestedProtoGeneric: (root: ProtoRoot, path: string[]) => any;
 export declare const getNested: (root: ProtoRoot, path: string[]) => any;
@@ -23,3 +23,25 @@ export declare const makeHookKeyName: (name: string) => any;
  * @returns Array of query method names match the given pattern.
  */
 export declare const getQueryMethodNames: (packagePath: string, methodKeys: string[], patterns?: string[]) => string[];
+/**
+ * test if a proto ref is included by the operation.
+ * @param ref a ProtoRef with proto file info and package.
+ * @param include patterns(will be deprecated soon), packages, proto files to include
+ * @returns
+ */
+export declare const isRefIncluded: (ref: ProtoRef, include?: {
+    patterns?: string[];
+    packages?: string[];
+    protos?: string[];
+}) => boolean;
+/**
+ * test if a proto ref is excluded from the operation.
+ * @param ref a ProtoRef with proto file info and package.
+ * @param exclude patterns(will be deprecated soon), packages, proto files to exclude
+ * @returns
+ */
+export declare const isRefExcluded: (ref: ProtoRef, exclude?: {
+    packages?: string[];
+    protos?: string[];
+}) => boolean;
+export declare const getObjectName: (name: string, scope?: string[]) => string;
