+++ packages/utils/types/utils.d.ts
@@ -1,3 +1,4 @@
+import { TelescopeOptions, TelescopeOption, ProtoRef } from '@cosmology/types';
 /**
  * swap the key and value of the input object
  * @param input obj needs to swap
@@ -8,3 +9,5 @@ export declare const swapKeyValue: (input: {
 }) => {
     [key: string]: string;
 };
+export declare const getPluginValue: (optionName: TelescopeOption | string, currentPkg: string, options: TelescopeOptions) => any;
+export declare const getTypeNameFromFieldName: (name: string, importSrc: string, ref: ProtoRef) => string;
